% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{heatinginfo_15min}
\alias{heatinginfo_15min}
\title{Heating info for 15-min smart meter data.}
\format{
A data frame with the following of variables: 

\describe{
\item{\code{VID}}{An pseudonym for the household}
\item{\code{household_type}}{The housing type: *single family home* (detached house),
  *multi-family home* (multiple dwellings in one house), *semidetached house* and 
  *teraced house* (multiple houses in a row)}
\item{\code{heating_type}}{Type of the heating system, either *electric heating*, 
  *heat pump*, *heat pump and boiler*, or *other* (including gas, central heating in 
  a multi-family home)}
\item{\code{survey_WP_type}}{Type of the heat pump, when a heat pump is installed, 
  according to the survey response. Can be either *air*, *geothermal*, or *don't know*.}
\item{\code{survey_WP_age}}{The age of the heat pump according to the survey 
  response. Can be either *<10 years*, *10-20 years*, *20-30 years*, *>30 years*, 
  or *don't know*}
}
}
\usage{
heatinginfo_15min
}
\description{
Ground truth data on housing type and heating information for the 15-minute 
smart meter dataset *elcons_15min*. The data was collectes from customers 
of an electric utility company in Switzerland with a survey in 2018.
}
\keyword{datasets}
